fetch space player count client side
