+++ apps/client/src/pages/app/[id].tsx
@@ -1,4 +1,4 @@
-import { GetServerSideProps, InferGetServerSidePropsType } from "next";
+import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
 import Script from "next/script";
 import { useEffect, useMemo, useRef, useState } from "react";
 
@@ -10,20 +10,13 @@ import { useAppStore } from "../../app/store";
 import ChatBox from "../../app/ui/ChatBox";
 import LoadingScreen from "../../app/ui/LoadingScreen";
 import UserButton from "../../app/ui/UserButtons";
-import {
-  getPublicationProps,
-  PublicationProps,
-} from "../../client/lens/utils/getPublicationProps";
+import { getPublicationProps } from "../../client/lens/utils/getPublicationProps";
 import MetaTags from "../../home/MetaTags";
 
-interface Props extends PublicationProps {
-  id: string;
-}
-
-export const getServerSideProps: GetServerSideProps<Props> = async ({
+export const getServerSideProps = async ({
   res,
   query,
-}) => {
+}: GetServerSidePropsContext) => {
   res.setHeader(
     "Cache-Control",
     "public, s-maxage=60, stale-while-revalidate=600"
