+++ apps/client/src/pages/space/[id]/settings.tsx
@@ -1,25 +1,25 @@
 import { useHidePublicationMutation } from "@wired-labs/lens";
-import { GetServerSideProps, InferGetServerSidePropsType } from "next";
+import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 
 import { useLens } from "../../../client/lens/hooks/useLens";
-import { PublicationProps } from "../../../client/lens/utils/getPublicationProps";
+import { getPublicationProps } from "../../../client/lens/utils/getPublicationProps";
 import { trpc } from "../../../client/trpc";
 import { getNavbarLayout } from "../../../home/layouts/NavbarLayout/NavbarLayout";
-import { getSpaceLayoutProps } from "../../../home/layouts/SpaceLayout/getSpaceLayoutProps";
 import SpaceLayout from "../../../home/layouts/SpaceLayout/SpaceLayout";
 import Button from "../../../ui/Button";
 
-export const getServerSideProps: GetServerSideProps<
-  PublicationProps & { host: string; playerCount: number | null }
-> = async ({ res, query }) => {
+export const getServerSideProps = async ({
+  res,
+  query,
+}: GetServerSidePropsContext) => {
   res.setHeader(
     "Cache-Control",
     "public, s-maxage=60, stale-while-revalidate=604800"
   );
 
-  const props = await getSpaceLayoutProps(query.id as string);
+  const props = await getPublicationProps(query.id as string);
 
   return {
     props,
